Move index.js code to bin/validate.js
